requirements.txt
aiofiles==22.1.0
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.29.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.83.0
greenlet==1.1.3
h11==0.13.0
httplib2==0.20.4
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
orjson==3.8.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==4.21.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
psycopg2==2.9.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
Random-Word==1.0.10
RandomWords==0.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wonderwords==2.2.0
bcrypt==3.1.7
pdf2image==1.16.0
Pillow==9.2.0
pypdfium2==3.5.0 